Move getCMSFieldOptionsAndValidIds to shared UI
