+++ apps/ff-site/src/app/_utils/getCMSFieldOptionsAndValidIds.ts
@@ -1,26 +1,22 @@
 import {
-  getCMSFieldOptions,
-  getCollectionConfig,
-} from '@filecoin-foundation/utils/cmsConfigUtils'
-import { mapCMSOptionsToListboxFormat } from '@filecoin-foundation/utils/mapCMSOptionsToListboxFormat'
-import type { CMSCollectionName } from '@filecoin-foundation/utils/types/cmsConfig'
+  getCMSFieldOptionsAndValidIds as sharedGetCMSFieldOptionsAndValidIds,
+  type CMSFieldOptionsAndValidIdsParams as SharedCMSFieldOptionsAndValidIdsParams,
+} from '@filecoin-foundation/utils/getCMSFieldOptionsAndValidIds'
 
 import configData from '@/data/cmsConfigSchema.json'
 
-
-export type CMSFieldOptionsAndValidIdsParams = {
-  collectionName: CMSCollectionName
-  fieldName: string
-}
+export type CMSFieldOptionsAndValidIdsParams = Omit<
+  SharedCMSFieldOptionsAndValidIdsParams,
+  'configData'
+>
 
 export function getCMSFieldOptionsAndValidIds({
   collectionName,
   fieldName,
 }: CMSFieldOptionsAndValidIdsParams) {
-  const { fields } = getCollectionConfig(collectionName, configData)
-  const cmsOptions = getCMSFieldOptions(fields, fieldName)
-  const options = mapCMSOptionsToListboxFormat(cmsOptions)
-  const validIds = options.map((setting) => setting.id)
-
-  return { options, validIds }
+  return sharedGetCMSFieldOptionsAndValidIds({
+    collectionName,
+    fieldName,
+    configData,
+  })
 }
